fix: reorganize messaging fixtures
