+++ tests/messaging/__init__.py
@@ -1,0 +1,1 @@
+"""Messaging tests package."""
