+++ tests/messaging/fixtures/__init__.py
@@ -1,0 +1,1 @@
+"""Fixtures for messaging tests."""
